0.7.1

+++ app/Http/Controllers/API/ShowController.php
@@ -6,10 +6,11 @@ use KRLX\Show;
 use KRLX\User;
 use Validator;
 use Illuminate\Http\Request;
+use KRLX\Mail\ShowSubmitted;
 use KRLX\Rulesets\ShowRuleset;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Mail;
 use KRLX\Http\Controllers\Controller;
-use KRLX\Notifications\ShowSubmitted;
 use KRLX\Notifications\ShowInvitation;
 use KRLX\Http\Requests\ShowUpdateRequest;
 use KRLX\Notifications\NewUserShowInvitation;
@@ -170,32 +171,43 @@ class ShowController extends Controller
      */
     public function join(Request $request, Show $show)
     {
-        $request->validate(['token' => 'required|string']); // (1)
+        $request->validate(['token' => 'required|string']);
         try {
-            $data = decrypt($request->input('token')); // (2)
-            if (! is_array($data)) { // (3)
-                throw new DecryptException('The given token is not an array.');
-            } elseif (! array_key_exists('show', $data) or ! array_key_exists('user', $data)) { // (4)
-                throw new DecryptException('The token does not have the required components.');
-            } elseif ($data['user'] != $request->user()->email) { // (5)
-                throw new DecryptException('The token does not belong to you.');
-            } elseif ($data['show'] != $show->id) { // (6)
-                throw new DecryptException('The token does not belong to this show.');
-            }
+            $this->validateToken($request->input('token'), $request->user()->email, $show->id);
         } catch (DecryptException $e) {
             abort(400, 'The token is invalid.');
         }
 
-        // We now know that: (1) the token is present, (2) it is encrypted,
-        // (3) the decrypted form is an array, (4) it has the required
-        // components, (5) it belongs to the user, and (6) it belongs to the
-        // show being joined. This request is therefore authorized and we can
-        // now make the connection.
         $show->invitees()->detach($request->user()->id);
-        $show->hosts()->detach($request->user()->id);
-        $show->hosts()->attach($request->user()->id, ['accepted' => true]);
+        if (! $request->has('cancel')) {
+            $show->hosts()->detach($request->user()->id);
+            $show->hosts()->attach($request->user()->id, ['accepted' => true]);
+        }
 
         return $show;
+    }
+
+    /**
+     * Validate a join token.
+     *
+     * @param  string  $token
+     * @param  string  $email
+     * @param  string  $show
+     * @throws DecryptException
+     * @return void
+     */
+    private function validateToken(string $token, string $email, string $show)
+    {
+        $data = decrypt($token);
+        if (! is_array($data)) {
+            throw new DecryptException('The given token is not an array.');
+        } elseif (! array_key_exists('show', $data) or ! array_key_exists('user', $data)) {
+            throw new DecryptException('The token does not have the required components.');
+        } elseif ($data['user'] != $email) {
+            throw new DecryptException('The token does not belong to you.');
+        } elseif ($data['show'] != $show) {
+            throw new DecryptException('The token does not belong to this show.');
+        }
     }
 
     /**
@@ -214,7 +226,7 @@ class ShowController extends Controller
 
             Validator::make($show->toArray(), $rules)->validate();
 
-            $request->user()->notify(new ShowSubmitted($show));
+            Mail::to($show->hosts)->send(new ShowSubmitted($show));
         }
         $show->submitted = $request->input('submitted') ?? false;
         $show->save();
